requirements.txt
boto3
python-dotenv
aos-metadata-collector
